Don't include VM.h in JSObject.h https://bugs.webkit.org/show_bug.cgi?id=292828
Reviewed by Justin Michaud and Mark Lam.

Move enough inline functions from *Inlines.h files so that we can avoid including VM.h
in JSObject.h to speed up build time.

* Source/JavaScriptCore/builtins/BuiltinUtils.h:
* Source/JavaScriptCore/bytecode/ArrayProfile.h:
* Source/JavaScriptCore/bytecode/BytecodeIntrinsicRegistry.h:
* Source/JavaScriptCore/interpreter/CachedCall.h:
* Source/JavaScriptCore/jit/AssemblyHelpers.h:
* Source/JavaScriptCore/runtime/ArrayPrototypeInlines.h:
* Source/JavaScriptCore/runtime/CachedTypes.h:
* Source/JavaScriptCore/runtime/DirectArguments.h:
* Source/JavaScriptCore/runtime/Identifier.h:
(JSC::Identifier::add): Deleted.
* Source/JavaScriptCore/runtime/IdentifierInlines.h:
(JSC::Identifier::Identifier):
(JSC::Identifier::add):
* Source/JavaScriptCore/runtime/JSObject.h:
(JSC::JSObject::canHaveExistingOwnIndexedGetterSetterProperties): Deleted.
(JSC::JSObject::canHaveExistingOwnIndexedProperties const): Deleted.
(JSC::JSObject::initializeIndex): Deleted.
(JSC::JSObject::initializeIndexWithoutBarrier): Deleted.
(JSC::JSObject::setButterfly): Deleted.
(JSC::JSObject::nukeStructureAndSetButterfly): Deleted.
(JSC::JSObject::get const): Deleted.
(JSC::JSObject::getAs const): Deleted.
* Source/JavaScriptCore/runtime/JSObjectInlines.h:
(JSC::JSObject::setButterfly):
(JSC::JSObject::nukeStructureAndSetButterfly):
(JSC::JSObject::get const):
(JSC::JSObject::getAs const):
(JSC::JSObject::initializeIndex):
(JSC::JSObject::initializeIndexWithoutBarrier):
(JSC::JSObject::canHaveExistingOwnIndexedGetterSetterProperties):
(JSC::JSObject::canHaveExistingOwnIndexedProperties const):
* Source/JavaScriptCore/runtime/ObjectConstructor.h:
* Source/JavaScriptCore/runtime/Operations.h:
* Source/JavaScriptCore/runtime/PropertySlot.h:
* Source/JavaScriptCore/runtime/RegExp.h:
* Source/JavaScriptCore/runtime/RegExpInlines.h:
(JSC::RegExp::subspaceFor):
* Source/JavaScriptCore/runtime/Structure.h:
(JSC::Structure::subspaceFor): Deleted.
* Source/JavaScriptCore/runtime/StructureInlines.h:
(JSC::Structure::subspaceFor):
(JSC::Structure::finishCreation):
* Source/JavaScriptCore/runtime/StructureRareData.h:
* Source/JavaScriptCore/runtime/StructureRareDataInlines.h:
(JSC::StructureRareData::subspaceFor):
* Source/JavaScriptCore/runtime/SyntheticModuleRecord.cpp:
* Source/JavaScriptCore/runtime/SyntheticModuleRecord.h:
* Source/WebCore/bindings/js/JSDOMConvertStrings.h:
* Source/WebCore/bindings/js/WebCoreBuiltinNames.h:

Canonical link: https://commits.webkit.org/294804@main
