+++ Source/JavaScriptCore/runtime/JSObject.h
@@ -36,16 +36,12 @@
 #include "IndexingHeaderInlines.h"
 #include "JSCast.h"
 #include "MathCommon.h"
-#include "ObjectInitializationScope.h"
 #include "PropertySlot.h"
 #include "PropertyStorage.h"
 #include "PutDirectIndexMode.h"
 #include "PutPropertySlot.h"
 #include "Structure.h"
 #include "StructureTransitionTable.h"
-#include "VM.h"
-#include "JSString.h"
-#include "SparseArrayValueMap.h"
 #include <wtf/StdLibExtras.h>
 
 WTF_ALLOW_UNSAFE_BUFFER_USAGE_BEGIN
@@ -67,12 +63,15 @@ class Exception;
 class GetterSetter;
 class InternalFunction;
 class JSFunction;
+class JSString;
 class LLIntOffsetsExtractor;
 class MarkedBlock;
+class ObjectInitializationScope;
 class PropertyDescriptor;
 class PropertyNameArray;
 class Structure;
 class ThrowScope;
+class VM;
 struct HashTable;
 struct HashTableValue;
 
@@ -151,12 +150,12 @@ public:
         
     inline bool mayInterceptIndexedAccesses();
 
-    JSValue get(JSGlobalObject*, PropertyName) const;
-    JSValue get(JSGlobalObject*, unsigned propertyName) const;
+    inline JSValue get(JSGlobalObject*, PropertyName) const; // Defined in JSObjectInlines.h
+    inline JSValue get(JSGlobalObject*, unsigned propertyName) const; // Defined in JSObjectInlines.h
     JSValue get(JSGlobalObject*, uint64_t propertyName) const;
 
     template<typename T, typename PropertyNameType>
-    T getAs(JSGlobalObject*, PropertyNameType) const;
+    inline T getAs(JSGlobalObject*, PropertyNameType) const; // Defined in JSObjectInlines.h
 
     template<bool checkNullStructure = false>
     bool getPropertySlot(JSGlobalObject*, PropertyName, PropertySlot&);
@@ -208,58 +207,11 @@ public:
         return m_butterfly->vectorLength();
     }
     
-    bool canHaveExistingOwnIndexedGetterSetterProperties()
-    {
-        if (!hasIndexedProperties(indexingType()))
-            return false;
-
-        switch (indexingType()) {
-        case ALL_BLANK_INDEXING_TYPES:
-        case ALL_UNDECIDED_INDEXING_TYPES:
-        case ALL_INT32_INDEXING_TYPES:
-        case ALL_CONTIGUOUS_INDEXING_TYPES:
-        case ALL_DOUBLE_INDEXING_TYPES:
-            return false;
-        case ALL_ARRAY_STORAGE_INDEXING_TYPES: {
-            SparseArrayValueMap* map = m_butterfly->arrayStorage()->m_sparseMap.get();
-            if (!map)
-                return false;
-            return map->hasAnyKindOfGetterSetterProperties();
-        }
-        default:
-            RELEASE_ASSERT_NOT_REACHED();
-        }
-    }
+    inline bool canHaveExistingOwnIndexedGetterSetterProperties(); // Defined in RenderObjectInlines.h
 
     // This is only valid after using canPerformFastPropertyEnumerationCommon().
     // This code is not checking getOwnPropertySlot override etc.
-    unsigned canHaveExistingOwnIndexedProperties() const
-    {
-        if (!hasIndexedProperties(indexingType()))
-            return false;
-
-        switch (indexingType()) {
-        case ALL_BLANK_INDEXING_TYPES:
-        case ALL_UNDECIDED_INDEXING_TYPES:
-            return false;
-        case ALL_INT32_INDEXING_TYPES:
-        case ALL_CONTIGUOUS_INDEXING_TYPES:
-        case ALL_DOUBLE_INDEXING_TYPES:
-            return m_butterfly->publicLength();
-        case ALL_ARRAY_STORAGE_INDEXING_TYPES: {
-            ArrayStorage* storage = m_butterfly->arrayStorage();
-            unsigned usedVectorLength = std::min(storage->length(), storage->vectorLength());
-            if (usedVectorLength)
-                return true;
-            SparseArrayValueMap* map = storage->m_sparseMap.get();
-            if (!map)
-                return false;
-            return map->size();
-        }
-        default:
-            RELEASE_ASSERT_NOT_REACHED();
-        }
-    }
+    inline unsigned canHaveExistingOwnIndexedProperties() const; // Defined in RenderObjectInlines.h
 
     static bool putInlineForJSObject(JSCell*, JSGlobalObject*, PropertyName, JSValue, PutPropertySlot&);
     
@@ -587,111 +539,17 @@ public:
         }
     }
 
-    void initializeIndex(ObjectInitializationScope& scope, unsigned i, JSValue v)
-    {
-        initializeIndex(scope, i, v, indexingType());
-    }
+    inline void initializeIndex(ObjectInitializationScope&, unsigned, JSValue); // Defined in JSObjectInlines.h
 
     // NOTE: Clients of this method may call it more than once for any index, and this is supposed
     // to work.
-    ALWAYS_INLINE void initializeIndex(ObjectInitializationScope& scope, unsigned i, JSValue v, IndexingType indexingType)
-    {
-        VM& vm = scope.vm();
-        Butterfly* butterfly = m_butterfly.get();
-        switch (indexingType) {
-        case ALL_UNDECIDED_INDEXING_TYPES: {
-            setIndexQuicklyToUndecided(vm, i, v);
-            break;
-        }
-        case ALL_INT32_INDEXING_TYPES: {
-            ASSERT(i < butterfly->publicLength());
-            ASSERT(i < butterfly->vectorLength());
-            if (!v.isInt32()) {
-                convertInt32ToDoubleOrContiguousWhilePerformingSetIndex(vm, i, v);
-                break;
-            }
-            [[fallthrough]];
-        }
-        case ALL_CONTIGUOUS_INDEXING_TYPES: {
-            ASSERT(i < butterfly->publicLength());
-            ASSERT(i < butterfly->vectorLength());
-            butterfly->contiguous().at(this, i).set(vm, this, v);
-            break;
-        }
-        case ALL_DOUBLE_INDEXING_TYPES: {
-            ASSERT(i < butterfly->publicLength());
-            ASSERT(i < butterfly->vectorLength());
-            if (!v.isNumber()) {
-                convertDoubleToContiguousWhilePerformingSetIndex(vm, i, v);
-                return;
-            }
-            double value = v.asNumber();
-            if (value != value) {
-                convertDoubleToContiguousWhilePerformingSetIndex(vm, i, v);
-                return;
-            }
-            butterfly->contiguousDouble().at(this, i) = value;
-            break;
-        }
-        case ALL_ARRAY_STORAGE_INDEXING_TYPES: {
-            ArrayStorage* storage = butterfly->arrayStorage();
-            ASSERT(i < storage->length());
-            ASSERT(i < storage->m_numValuesInVector);
-            storage->m_vector[i].set(vm, this, v);
-            break;
-        }
-        default:
-            RELEASE_ASSERT_NOT_REACHED();
-        }
-    }
-        
-    void initializeIndexWithoutBarrier(ObjectInitializationScope& scope, unsigned i, JSValue v)
-    {
-        initializeIndexWithoutBarrier(scope, i, v, indexingType());
-    }
+    ALWAYS_INLINE void initializeIndex(ObjectInitializationScope&, unsigned, JSValue, IndexingType); // Defined in JSObjectInlines.h
+
+    inline void initializeIndexWithoutBarrier(ObjectInitializationScope&, unsigned, JSValue); // Defined in JSObjectInlines.h
 
     // This version of initializeIndex is for cases where you know that you will not need any
     // barriers. This implies not having any data format conversions.
-    ALWAYS_INLINE void initializeIndexWithoutBarrier(ObjectInitializationScope&, unsigned i, JSValue v, IndexingType indexingType)
-    {
-        Butterfly* butterfly = m_butterfly.get();
-        switch (indexingType) {
-        case ALL_UNDECIDED_INDEXING_TYPES: {
-            RELEASE_ASSERT_NOT_REACHED();
-            break;
-        }
-        case ALL_INT32_INDEXING_TYPES: {
-            ASSERT(i < butterfly->publicLength());
-            ASSERT(i < butterfly->vectorLength());
-            RELEASE_ASSERT(v.isInt32());
-            [[fallthrough]];
-        }
-        case ALL_CONTIGUOUS_INDEXING_TYPES: {
-            ASSERT(i < butterfly->publicLength());
-            ASSERT(i < butterfly->vectorLength());
-            butterfly->contiguous().at(this, i).setWithoutWriteBarrier(v);
-            break;
-        }
-        case ALL_DOUBLE_INDEXING_TYPES: {
-            ASSERT(i < butterfly->publicLength());
-            ASSERT(i < butterfly->vectorLength());
-            RELEASE_ASSERT(v.isNumber());
-            double value = v.asNumber();
-            RELEASE_ASSERT(value == value);
-            butterfly->contiguousDouble().at(this, i) = value;
-            break;
-        }
-        case ALL_ARRAY_STORAGE_INDEXING_TYPES: {
-            ArrayStorage* storage = butterfly->arrayStorage();
-            ASSERT(i < storage->length());
-            ASSERT(i < storage->m_numValuesInVector);
-            storage->m_vector[i].setWithoutWriteBarrier(v);
-            break;
-        }
-        default:
-            RELEASE_ASSERT_NOT_REACHED();
-        }
-    }
+    ALWAYS_INLINE void initializeIndexWithoutBarrier(ObjectInitializationScope&, unsigned, JSValue, IndexingType); // Defined in JSObjectInlines.h
         
     bool hasSparseMap()
     {
@@ -944,11 +802,11 @@ public:
     JS_EXPORT_PRIVATE Butterfly* allocateMoreOutOfLineStorage(VM&, size_t oldSize, size_t newSize);
 
     // Call this when you do not need to change the structure.
-    void setButterfly(VM&, Butterfly*);
-    
+    inline void setButterfly(VM&, Butterfly*); // Defined in JSObjectInlines.h
+
     // Call this if you do need to change the structure, or if you changed something about a structure
     // in-place.
-    void nukeStructureAndSetButterfly(VM&, StructureID oldStructureID, Butterfly*);
+    inline void nukeStructureAndSetButterfly(VM&, StructureID oldStructureID, Butterfly*); // Defined in JSObjectInlines.h
 
     void setStructure(VM&, Structure*);
 
@@ -1410,31 +1268,6 @@ inline void JSObject::setStructure(VM& vm, Structure* structure)
     JSCell::setStructure(vm, structure);
 }
 
-inline void JSObject::setButterfly(VM& vm, Butterfly* butterfly)
-{
-    if (isX86() || vm.heap.mutatorShouldBeFenced()) {
-        WTF::storeStoreFence();
-        m_butterfly.set(vm, this, butterfly);
-        WTF::storeStoreFence();
-        return;
-    }
-
-    m_butterfly.set(vm, this, butterfly);
-}
-
-inline void JSObject::nukeStructureAndSetButterfly(VM& vm, StructureID oldStructureID, Butterfly* butterfly)
-{
-    if (isX86() || vm.heap.mutatorShouldBeFenced()) {
-        setStructureIDDirectly(oldStructureID.nuke());
-        WTF::storeStoreFence();
-        m_butterfly.set(vm, this, butterfly);
-        WTF::storeStoreFence();
-        return;
-    }
-
-    m_butterfly.set(vm, this, butterfly);
-}
-
 inline JSObject* asObject(JSCell* cell)
 {
     ASSERT(cell);
@@ -1610,50 +1443,6 @@ ALWAYS_INLINE bool JSObject::getPropertySlot(JSGlobalObject* globalObject, Prope
     if (std::optional<uint32_t> index = parseIndex(propertyName))
         return getPropertySlot(globalObject, index.value(), slot);
     return false;
-}
-
-inline JSValue JSObject::get(JSGlobalObject* globalObject, PropertyName propertyName) const
-{
-    VM& vm = getVM(globalObject);
-    auto scope = DECLARE_THROW_SCOPE(vm);
-    PropertySlot slot(this, PropertySlot::InternalMethodType::Get);
-    bool hasProperty = const_cast<JSObject*>(this)->getPropertySlot(globalObject, propertyName, slot);
-
-    EXCEPTION_ASSERT(!scope.exception() || vm.hasPendingTerminationException() || !hasProperty);
-    RETURN_IF_EXCEPTION(scope, jsUndefined());
-
-    if (hasProperty)
-        RELEASE_AND_RETURN(scope, slot.getValue(globalObject, propertyName));
-
-    return jsUndefined();
-}
-
-inline JSValue JSObject::get(JSGlobalObject* globalObject, unsigned propertyName) const
-{
-    VM& vm = getVM(globalObject);
-    auto scope = DECLARE_THROW_SCOPE(vm);
-    PropertySlot slot(this, PropertySlot::InternalMethodType::Get);
-    bool hasProperty = const_cast<JSObject*>(this)->getPropertySlot(globalObject, propertyName, slot);
-
-    EXCEPTION_ASSERT(!scope.exception() || vm.hasPendingTerminationException() || !hasProperty);
-    RETURN_IF_EXCEPTION(scope, jsUndefined());
-
-    if (hasProperty)
-        RELEASE_AND_RETURN(scope, slot.getValue(globalObject, propertyName));
-
-    return jsUndefined();
-}
-
-template<typename T, typename PropertyNameType>
-inline T JSObject::getAs(JSGlobalObject* globalObject, PropertyNameType propertyName) const
-{
-    JSValue value = get(globalObject, propertyName);
-#if ASSERT_ENABLED || ENABLE(SECURITY_ASSERTIONS)
-    VM& vm = getVM(globalObject);
-    if (vm.exceptionForInspection())
-        return nullptr;
-#endif
-    return jsCast<T>(value);
 }
 
 inline bool JSObject::putDirect(VM& vm, PropertyName propertyName, JSValue value, unsigned attributes)
