+++ Source/JavaScriptCore/runtime/PropertySlot.h
@@ -24,14 +24,17 @@
 #include "DisallowVMEntry.h"
 #include "GetVM.h"
 #include "JSCJSValue.h"
+#include "JSCPtrTag.h"
 #include "PropertyName.h"
 #include "PropertyOffset.h"
 #include "ScopeOffset.h"
+#include "Watchpoint.h"
 #include <wtf/Assertions.h>
 #include <wtf/ForbidHeapAllocation.h>
 #include <wtf/FunctionPtr.h>
 
 namespace JSC {
+
 class GetterSetter;
 class JSObject;
 class JSModuleEnvironment;
